client/requirements.txt
astroid==2.5.1
grpcio==1.36.1
grpcio-tools==1.36.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pkg-resources==0.0.0
protobuf==3.15.6
pylint==2.7.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
